Fix handling of `{ port: '0' }`
Being passed `{ port: '0' }` (`{ port: string }`) is a Typescript type
error, but we explicitly have a test that checks that it is gracefully
handled, with behavior matching being passed

- `{ port: 0 }` (`{ port: number }`),
- `'0'` (`string`), or
- `0` (`number`),

which all convert the `port` to `0`.

However, that test was actually dependent on the following leak from a
test above it:

    process.env.HOST = 'localhost';

Switching to `vi.stubEnv` and ensuring to call `vi.unstubAllEnvs`
reveals that the behavior is broken.

This updates the behavior to apply the same check to `_userOptions.port`
as we do to `_userOptions`, and coerce `number` and `string` values to
integers.